podział na pakiety, podział klasy głównej
