+++ kodilla-individual-project/src/main/java/com/kodilla/game/gameLogic/Sign.java
@@ -1,4 +1,4 @@
-package com.kodilla.game;
+package com.kodilla.game.gameLogic;
 
 public enum Sign {
     CIRCLE,
