+++ kodilla-individual-project/src/test/java/GameStateTestSuite.java
@@ -1,4 +1,6 @@
-import com.kodilla.game.*;
+import com.kodilla.game.gameLogic.GameResult;
+import com.kodilla.game.gameLogic.GameState;
+import com.kodilla.game.gameLogic.Sign;
 import org.junit.*;
 
 import java.util.HashMap;
